FIX: Make sure CC addresses for forwarded emails are used
When forwarding emails into the group inbox we are currently
ignoring the original CC addresses on the forwarded email. This
commit makes sure we record those CC addresses on the incoming
email record for the topic, as well as inviting those CC addresses
as staged users on the topic when it is created.