UX: improve error message when no trust level to login from DiscourseHub
DiscourseHub requires generating API keys to login, which is affected by
the min_trust_level_for_user_api_key site setting.

If it is increased over the default of 0, then new users with 0 trust
level will not be able to login with the app, and in that case we would
be showing them a technical error message involving API keys, which
might be bad UX for newcomers to Discourse.

This commit differentiates exactly that case, and serves a less
technical error message.